requirements.txt
flet
numpy
Pillow
scikit-learn
joblib
